% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bypass_autoinit.R
\name{bypass_reticulate_autoinit}
\alias{bypass_reticulate_autoinit}
\title{Prevent Reticulate Autoconfiguring Python environments}
\usage{
bypass_reticulate_autoinit()
}
\description{
Adds 'RETICULATE_AUTOCONFIGURE = "FALSE"' to .Renviron to prevent 'reticulate' automatically loading Python environments.
}
