% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CODrecovery}
\alias{CODrecovery}
\title{Three Recovering Cod Populations}
\format{
A list of three dataframes with 191 rows and 6 variables each:
\describe{
  \item{community_id}{the identity of the simulated community}
  \item{time}{time index}
  \item{biomass}{population total biomass}
  \item{mean.size}{average length of cod individuals}
  \item{sd.size}{variation in length of cod individuals}
  \item{inflection_pt}{the time index where transition occurs}
}
}
\source{
Clements C., McCarthy M., Blanchard J. (2019) Early warning signals of recovery in complex systems. Nature Communications, 10:1681.
}
\usage{
CODrecovery
}
\description{
A dataset containing three simulated cod populations. Community 1 does not recovery whereas Community 100 and 200 do.
}
\examples{
data("CODrecovery", package = "EWSmethods")

cod_data <- CODrecovery$scenario2

}
\keyword{datasets}
