% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_composite_ews_helper_fns.R
\name{CV}
\alias{CV}
\title{Coefficient of Variation}
\usage{
CV(x, na.rm)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Boolean. If TRUE, missing values are removed.}
}
\description{
Coefficient of Variation
}
