% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_sewsnet.R
\name{default_sewsnet_path}
\alias{default_sewsnet_path}
\title{Path to S-EWSNet Model}
\usage{
default_sewsnet_path()
}
\value{
No return value, called for reference.
}
\description{
The default path to the S-EWSNet model. Is the location of the EWSmethods package.
}
