% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_ewsnet.R
\name{default_weights_path}
\alias{default_weights_path}
\title{Path to Model Weights}
\usage{
default_weights_path()
}
\value{
No return value, called for reference.
}
\description{
The default path for EWSNet model weights to use. Is the location of the EWSmethods package. If you'd like to instead set your own path, \code{ewsnet_reset()} contains the argument \code{weights_path} to do so.
}
