% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewsnet_init.R
\name{EWSNET_init}
\alias{EWSNET_init}
\title{EWSNET Initialisation}
\usage{
EWSNET_init(envname)
}
\arguments{
\item{envname}{A string naming the desired Python environment to create/activate. If no python or environment found, the functions prompts to install Anaconda and the required python packages}
}
\description{
EWSNET Initialisation
}
