% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_ewsnet.R
\name{ewsnet_reset}
\alias{ewsnet_reset}
\title{Reset EWSNet Model Weights}
\usage{
ewsnet_reset(remove_weights = FALSE, auto = FALSE)
}
\arguments{
\item{remove_weights}{Boolean. Should all weights be removed (\code{TRUE}) or should weights be re/downloaded (\code{FALSE}).}

\item{auto}{Boolean. If \code{TRUE}, no user confirmation is required for re/download.}
}
\description{
Restores EWSNet model weights to the pretrained defaults published at https://ewsnet.github.io. This is vital on first use of EWSNet as no model weights are provided with `EWSmethods`. The use of this function may also be necessary after finetuning to reset the model.
}
\examples{
\dontrun{
ewsnet_reset(remove_weights = FALSE, auto = FALSE) # on first use of EWSNet via `EWSmethods`
}

\dontrun{
ewsnet_reset(remove_weights = TRUE) # to remove all downloaded weights
}

}
