% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewsnet_predict.R
\name{ewsNETw_25}
\alias{ewsNETw_25}
\title{EWSNET Predict}
\usage{
ewsNETw_25(x)
}
\arguments{
\item{x}{A numeric vector of values to be tested}
}
\value{
A data.frame of EWSNET predictions
}
\description{
EWSNET Predict
}
