% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFisher.R
\name{GFisher}
\alias{GFisher}
\title{Calculate Fisher Information}
\usage{
GFisher(timedat, data, sost, hwin, winspace = 1, TL = 90)
}
\arguments{
\item{timedat}{A numeric vector of equal spacing representing the time points}

\item{data}{A numeric matrix of individual time series across the columns. These could be different species, populations or measurements.}

\item{sost}{A 1 x n matrix where n is a length equal to the number of time series in \code{data}. Each value is the 'size of state' tolerable for that time series and typically is represented by the standard deviation of the time series during a reference period.}

\item{hwin}{Numeric value. The number of data points within the window.}

\item{winspace}{Numeric value. The number of data points to roll the window over in each iteration. Must be less than \code{hwin}.}

\item{TL}{Numeric value. The 'tightening level' or percentage of points shared between states that allows the algorithm to classify data points as the same state.}
}
\value{
A list containing three objects:
\item{FI}{A numeric vector of Fisher information estimates}
\item{midt_win}{A numeric vector of the time index at the centre of the window for that associated value in \code{FI}.}
\item{t_win}{A n x m numeric matrix where the length of n is the winspace and length of m is the number of window shifts made. Values are consequently the timepoint indices that contribute to that window.}
}
\description{
Uses a multivariate array of time series to estimate Fisher information following the approach of Karunanithi et al. (2010).
}
