% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_composite_ews_helper_fns.R
\name{interp}
\alias{interp}
\title{Linear Interpolation}
\usage{
interp(days, obs)
}
\arguments{
\item{days}{A numeric vector.}

\item{obs}{A numeric vector.}
}
\description{
Linear Interpolation
}
