% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_resilience_indicators.R
\name{maf}
\alias{maf}
\title{Maximum/Minimum Autocorrelation Factors}
\usage{
maf(x)
}
\arguments{
\item{x}{dataframe A n x m dataframe species (columns) abundances through time (rows)}
}
\value{
out A list containing the input data, the dimension reduction time series, rotations, autocorrelation and eigenvalues
}
\description{
Maximum/Minimum Autocorrelation Factors
}
