% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_composite_ews_wrapper.R
\name{multivariate_EWS_wrapper}
\alias{multivariate_EWS_wrapper}
\title{Perform Multivariate Early Warning Signal Assessment}
\usage{
multivariate_EWS_wrapper(
  data,
  metrics,
  threshold = 2,
  tail.direction = "one.tailed",
  burn_in = 5,
  ggplotIt = T,
  y_lab = "Generic Indicator Name",
  interpolate = F,
  method = c("expanding", "rolling"),
  winsize = 50
)
}
\arguments{
\item{data}{A dataframe where first column is time (equally spaced) and second column is abundance.}

\item{metrics}{String vector of early warning signal metrics. Options = c("cv", "acf", "ar1", "dr", "rr", "skew","kurt","mean.size","sd.size", "size.95","SD","trait").}

\item{threshold}{= "1" or "2". Threshold*sigma is the value which if EWS strength exceeds constitutes a "signal".}

\item{tail.direction}{= "one.tailed" or "two.tailed"."one.tailed" only indicates a warning if positive threshold sigma exceeded, "two.tailed"  indicates a warning if positive OR negative threshold*sigma exceeded.}

\item{burn_in}{= number of data points to 'train' signals prior to EWS assessment.}

\item{ggplotIt}{= ggplot plot of EWS strength trends AND input abundance.}

\item{y_lab}{= if ggplotIt = TRUE, labels abundance y axis.}

\item{interpolate}{= TRUE interpolates missing values found within the abundance time series.}

\item{method}{= c("expanding","rolling")."expanding" calls composite, expanding window EWS assessment. "rolling" calls typical, rolling window EWS assessment.}

\item{trait}{A vector of trait values if desired.}

\item{plotIt}{= base R plot of EWS strength trends.}

\item{trait_lab}{= if ggplotIt = TRUE, & trait populated, & "trait" supplied in metrics, labels abundance second abundance y axis (represents trait values through time).}

\item{trait_scale}{= scales trait y axis relative to abundance y axis.}

\item{data_source}{= NULL. A string to categorise result. Useful if performing multiple assessments iteratively.}
}
\value{
If ggplotIt = F, returns just EWS output.If ggplotIt = T, returns EWS output and plot object
}
\description{
Perform Multivariate Early Warning Signal Assessment
}
