% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFisherpdf.R
\name{NFisherpdf}
\alias{NFisherpdf}
\title{Fisher Information Binning}
\usage{
NFisherpdf(data, lmin, lmax, sizeofstates, TL)
}
\arguments{
\item{data}{A numeric matrix of individual time series across the columns. These could be different species, populations or measurements.}

\item{lmin}{Numeric value. Index of lower limit of window.}

\item{lmax}{Numeric value. Index of upper limit of window.}

\item{sizeofstates}{A 1 x n matrix where n is a length equal to the number of time series in \code{data}. Each value is the 'size of state' tolerable for that time series and typically is represented by the standard deviation of the time series during a reference period.}

\item{TL}{Numeric value. The 'tightening level' or percentage of points shared between states that allows the algorithm to classify data points as the same state.}
}
\value{
A list \code{NewList} containing to two objects:
\item{pdf}{A uniform probability density function.}
\item{neighbour}{The time point indices that contribute to the calculation.}
}
\description{
The purpose of this program is to bin the points into states by calculating the difference (dist) of each variable vector per timestep and then counting the variable vector if dist<= size of the state AND the percentage of variables in the vector that meets the criteria is over the tightening level.
}
\keyword{internal}
