% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_plot_rolling_ews_fn.R
\name{no.plot.ews}
\alias{no.plot.ews}
\title{Rolling Window Early Warning Signals}
\usage{
no.plot.ews(
  timeseries,
  winsize = 50,
  detrending = c("no", "gaussian", "loess", "linear", "first-diff"),
  bandwidth = NULL,
  span = NULL,
  degree = NULL,
  logtransform = FALSE,
  interpolate = FALSE,
  AR_n = FALSE,
  powerspectrum = FALSE
)
}
\description{
Rolling Window Early Warning Signals
}
