% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_ewsnet.R
\name{reset_ewsnet}
\alias{reset_ewsnet}
\title{Reset EWSNet Model Weights}
\usage{
reset_ewsnet(new_weight_location)
}
\arguments{
\item{new_weight_location}{A path string to the location where the unzipped file has been saved. Include "Pretrained" as the final path folder.}
}
\description{
Restores EWSNet model weights to the pretrained defaults published at https://ewsnet.github.io. This may be necessary after finetuning to reset the model. Follow this link (https://github.com/sahilsid/ewsnet) to find the original weights and extract the zip file in to a suitable location to use in \code{new_weight_location}.
}
\examples{
\dontrun{
reset_ewsnet(new_weight_location = "C:/Users/your_name/Downloads/Pretrained") # Windows example
}

\dontrun{
reset_ewsnet(new_weight_location = "/Users/your_name/Downloads/Pretrained") # Mac example
}

}
