% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundTO.R
\name{roundTO}
\alias{roundTO}
\title{Round 2 Function}
\usage{
roundTO(x, y)
}
\arguments{
\item{x}{Numeric value to be rounded.}

\item{y}{Numeric value to round \code{x} to nearest multiple of.}
}
\value{
Numeric value.
}
\description{
Rounding function renamed 'roundTO' to prevent homonyms
}
