% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smapJ_index.R
\name{smapJI}
\alias{smapJI}
\title{S-map Jacobian index function}
\source{
Ushio, M., Hsieh, Ch., Masuda, R. et al. (2018) Fluctuating interaction network and time-varying stability of a natural fish community. Nature 554, 360–363.
}
\usage{
smapJI(data, winsize = 50, theta_seq = NULL)
}
\arguments{
\item{data}{Numeric matrix with time in first column and species abundances in other columns}

\item{winsize}{Numeric. Defines the window size of the rolling window as a percentage of the time series length.}

\item{theta_seq}{Numeric vector of thetas (nonlinear tuning parameters) to estimate the Jacobian over. If `NULL`, a default sequence is provided.}
}
\value{
A dataframe where the first column is last time index of the window and the second column is the estimated index value. A value <1.0 indicates stability, a value >1.0 indicates instability.
}
\description{
Calculate a stability metric from the s-map estimated Jacobian
}
\examples{
#Load the multivariate simulated
#dataset `simTransComms`

data(simTransComms)

#Subset the "1_38_1" community prior to the transition

pre_simTransComms <- subset(simTransComms$`1_38_1`,time < inflection_pt)

#Estimate the stability index for the "1_38_1" community

egJI <- smapJI(data = pre_simTransComms[,2:7],
winsize = 25)

}
