% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_composite_ews_fn.R
\name{W_composite_ews}
\alias{W_composite_ews}
\title{Expanding Window Early Warning Signals}
\usage{
W_composite_ews(
  dat,
  indicators,
  weights,
  trait = NULL,
  threshold = 2,
  burn_in = 5,
  tail.direction = "one.tailed",
  interpolate = F
)
}
\arguments{
\item{dat}{A n x m dataframe with the first column  time indices and second column abundances.}

\item{indicators}{A string vector of the early warning signal indicators of interest. Options include: \code{"ar1"}, \code{"cv"}, \code{"SD"}, \code{"acf"},\code{"rr"},\code{"dr"},\code{"skew"},\code{"kurt"},\code{"mean.size"},\code{"sd.size"},\code{"sd.95"},\code{"trait"}.}

\item{weights}{A two column dataframe of weights for each indicator. First column is the indicators and second the weights.}

\item{trait}{Boolean. If true, \code{"trait"} has been provided as an indicator.}

\item{threshold}{Numeric. The threshold*sigma cutoff past which a warning is identified.}

\item{burn_in}{Numeric. Number of data points used to train signals before early warning signal assessment is performed.}

\item{tail.direction}{A string match to \code{"one.tailed"} or \code{"two.tailed"}.}

\item{interpolate}{Boolean. Should missing values in dat be interpolated.}
}
\value{
output A matrix of early warning signal indicators through time.
}
\description{
Modified code from Clements, McCarthy, and Blanchard (2019) Nature
Coms - "Early warning signals of rocovery in complex systems". Additional
EWS indicators, burn_in and interpolation capability added
}
\keyword{internal}
