% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_resilience_indicators.R
\name{wMAF}
\alias{wMAF}
\title{Multivariate Early Warning Signals}
\usage{
wMAF(
  data,
  method = c("rolling", "expanding"),
  winsize,
  burn_in = 5,
  tail.direction = "one.tailed",
  threshold = 2
)
}
\arguments{
\item{data}{A n x m dataframe with the first column time indices and remainder of columns species abundances.}

\item{method}{c("expanding","rolling")."expanding" calls composite, expanding window EWS assessment. "rolling" calls typical, rolling window EWS assessment.}

\item{winsize}{Numeric. If "method" = "rolling",defines the window size of the rolling window.}

\item{burn_in}{Numeric. If "method" = "expanding, defines the number of data points to 'train' signals prior to EWS assessment.}

\item{tail.direction}{A string. If "method" = "expanding, should both positive and negative thresholds be considered.}

\item{threshold}{Numeric. If "method" = "expanding, defines the threshold*sigma warning threshold.}
}
\value{
A list containing "raw" (the early warning signals through time) and "dimred.ts" (the dimension reduction time series)
}
\description{
Multivariate Early Warning Signals
}
