% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewsnet_init.R
\name{ewsnet_init}
\alias{ewsnet_init}
\title{EWSNet Initialisation}
\usage{
ewsnet_init(envname, pip_ignore_installed = FALSE)
}
\arguments{
\item{envname}{A string naming the desired Python environment to create/activate. If no python or environment found, the functions prompts to install Anaconda and the required python packages.}

\item{pip_ignore_installed}{Boolean. If \code{FALSE}, any packages already installed are loaded and not re-downloaded. However, if \code{TRUE}, packages are downloaded irregardless, overwriting any version already present (is analagous to updating if required).}
}
\value{
Does not return an object as is simply preparing your R session.
}
\description{
Prepares your R session for communicating with Python. This function first searches your computer for an appropriate Python environment and activates it, importing the vital Python packages required. If no appropriate Python install or environment is found, after asking permission, Anaconda is downloaded and an environment created.
}
\examples{
\dontrun{bypass_autoinit()}
#this function prevents `reticulate` automatically
#reloading the python environment from the
#previous session. Only needs to be run once
#on first install of `EWSmethods`

\dontrun{ewsnet_init(envname = "EWSNET_env")}

\dontrun{reticulate::py_config()}
#If successful, 'EWSNET_env forced by use_python
#function' will be printed.

}
