% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewsnet_predict.R
\name{ewsNETpredict}
\alias{ewsNETpredict}
\title{EWSNET Predict}
\usage{
ewsNETpredict(x, noise_type = "W", ensemble = 25)
}
\arguments{
\item{x}{A numeric vector of values to be tested.}

\item{noise_type}{A string stating the form of noise to use. Options are "W" or "C".}

\item{ensemble}{A numeric value stating the number of models to average over. Options range from 1 to 25.}
}
\value{
A dataframe of EWSNET predictions.
}
\description{
Communicates with EWSNet (https://ewsnet.github.io), a deep learning framework for modeling and anticipating regime shifts in dynamical systems, and returns the model's prediction for your univariate time series.
}
\examples{
abundance_data <- data.frame(time = seq(1:50), abundance = rnorm(50,mean = 20))
\dontrun{pred<- ewsNETpredict(abundance_data$abundance, noise_type = "W", ensemble = 15)}

}
