% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewsnet_predict.R
\name{ewsNETpredict}
\alias{ewsNETpredict}
\title{EWSNET Predict}
\usage{
ewsNETpredict(x, noise_type = "W", ensemble = "25")
}
\arguments{
\item{x}{A string stating the number of models to average over. Options range from 1 to 25.}
}
\value{
A data.frame of EWSNET predictions.
}
\description{
EWSNET Predict
}
