% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_plot_rolling_ews_fn.R
\name{no.plot.ews}
\alias{no.plot.ews}
\title{Rolling Window Early Warning Signals
An adaptation of the \code{generic_ews} function in the R package \code{earlywarnings}.}
\usage{
no.plot.ews(
  timeseries,
  winsize = 50,
  detrending = c("no", "gaussian", "loess", "linear", "first-diff"),
  bandwidth = NULL,
  span = NULL,
  degree = NULL,
  logtransform = FALSE,
  interpolate = FALSE,
  AR_n = FALSE,
  powerspectrum = FALSE
)
}
\arguments{
\item{timeseries}{A dataframe where first column is time (equally spaced) and second column is abundance.}

\item{winsize}{Numeric. Defines the window size of the rolling window as a percentage of the timeseries.}

\item{detrending}{A string. Should detrending be performed on the time series.}

\item{bandwidth}{Numeric. Bandwidth used for the Gaussian kernel when gaussian filtering is applied. It is expressed as percentage of the time series length.}

\item{span}{Numeric. Parameter that controls the degree of smoothing (numeric between 0 and 100, Default 25).}

\item{degree}{Numeric. The degree of polynomial when loess detrending is applied.}

\item{logtransform}{Boolean. If TRUE, data is log transformed as log(x+1).}

\item{interpolate}{Boolean. If TRUE, interpolates missing values found within the abundance time series.}

\item{AR_n}{Boolean. If TRUE, the best fitted AR(n) model is fitted to the data.}

\item{powerspectrum}{Boolean. If TRUE the power spectrum within each rolling window.}
}
\description{
Rolling Window Early Warning Signals
An adaptation of the \code{generic_ews} function in the R package \code{earlywarnings}.
}
\keyword{internal}
