% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_resilience_indicators.R
\name{wMAF}
\alias{wMAF}
\title{Multivariate Early Warning Signals}
\usage{
wMAF(
  data,
  method = c("rolling", "expanding"),
  winsize,
  burn_in = 5,
  tail.direction = "one.tailed",
  threshold = 2
)
}
\arguments{
\item{method}{c("expanding","rolling")."expanding" calls composite, expanding window EWS assessment. "rolling" calls typical, rolling window EWS assessment.}

\item{winsize}{Numeric. If \code{"method" = "rolling"},defines the window size of the rolling window.}

\item{burn_in}{Numeric. If \code{"method" = "expanding}, defines the number of data points to 'train' signals prior to EWS assessment.}

\item{tail.direction}{A string. If \code{"method" = "expanding}, should both postive and negative thresholds be considered.}

\item{threshold}{Numeric. If \code{"method" = "expanding}, defines the threshold*sigma warning threshold.}

\item{dat}{A n x m dataframe with the first column time indices and remainder of columns species abundances.}
}
\value{
A list containing \code{"raw"} (the early warning signals through time) and \code{"dimred.ts"} (the dimension reduction time series)
}
\description{
Multivariate Early Warning Signals
}
