% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_resilience_indicators.R
\name{wMAF}
\alias{wMAF}
\title{Multivariate Early Warning Signals}
\usage{
wMAF(
  data,
  method = c("rolling", "expanding"),
  winsize,
  burn_in = 5,
  tail.direction = "one.tailed",
  threshold = 2
)
}
\arguments{
\item{data}{A n x m dataframe with the first column time indices with the remainder of the columns species abundances.}

\item{method}{A single string stating either \code{"expanding"} or \code{"rolling"}.\code{"expanding"} calls the composite, expanding window EWS assessment. \code{"rolling"} calls the typical, rolling window EWS assessment.}

\item{winsize}{Numeric. If "method" = \code{"rolling"}, defines the window size of the rolling window as a percentage of the time series length.}

\item{burn_in}{Numeric. If "method" = \code{"expanding}, defines the number of data points to 'train' signals prior to EWS assessment.}

\item{tail.direction}{A string. If "method" = \code{"expanding}, should both positive and negative thresholds be considered.}

\item{threshold}{Numeric. If "method" = \code{"expanding}, defines the threshold*sigma warning threshold.}
}
\value{
A list containing \code{"raw"} (the early warning signals through time) and \code{"dimred.ts"} (the dimension reduction time series)
}
\description{
Multivariate Early Warning Signals
}
